api/pyproject.toml
[project]
name = "rootops-api"
version = "1.0.0"
description = "The AI-Native Internal Developer Platform API"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "pgvector==0.3.6",
    "pydantic-settings==2.7.1",
    "alembic==1.14.1",
    "python-dotenv==1.0.1",
    "sentence-transformers>=3.0.0",
    "transformers>=4.40.0,<5.0.0",
    "gitpython>=3.1.0",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "anthropic>=0.25.0",
    "PyGithub>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
